backend/requirements.txt
fastapi
uvicorn[standard]
pymongo
requests
beautifulsoup4
readability-lxml
python-dotenv
transformers
torch